fix: moved variablesservice and patternstore
