+++ backend/src/main/java/org/eclipse/tractusx/puris/backend/common/util/PatternStore.java
@@ -18,7 +18,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.eclipse.tractusx.puris.backend.common.api.logic.service;
+package org.eclipse.tractusx.puris.backend.common.util;
 
 import java.util.regex.Pattern;
