+++ backend/src/main/java/org/eclipse/tractusx/puris/backend/masterdata/domain/model/Site.java
@@ -27,7 +27,7 @@ import jakarta.persistence.Id;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Pattern;
 import lombok.*;
-import org.eclipse.tractusx.puris.backend.common.api.logic.service.PatternStore;
+import org.eclipse.tractusx.puris.backend.common.util.PatternStore;
 
 import java.util.SortedSet;
 import java.util.TreeSet;
