+++ backend/src/main/java/org/eclipse/tractusx/puris/backend/stock/logic/dto/FrontendMaterialDto.java
@@ -26,7 +26,7 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import org.eclipse.tractusx.puris.backend.common.api.logic.service.PatternStore;
+import org.eclipse.tractusx.puris.backend.common.util.PatternStore;
 
 @Getter
 @Setter
