+++ backend/src/main/java/org/eclipse/tractusx/puris/backend/stock/logic/dto/ProductStockRequestForMaterialDto.java
@@ -25,7 +25,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Pattern;
 import lombok.*;
-import org.eclipse.tractusx.puris.backend.common.api.logic.service.PatternStore;
+import org.eclipse.tractusx.puris.backend.common.util.PatternStore;
 
 /**
  * One object that is received as a response.
