+++ backend/src/main/java/org/eclipse/tractusx/puris/backend/stock/logic/dto/itemstocksamm/OrderPositionReference.java
@@ -26,7 +26,7 @@ import jakarta.validation.constraints.Pattern;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
-import org.eclipse.tractusx.puris.backend.common.api.logic.service.PatternStore;
+import org.eclipse.tractusx.puris.backend.common.util.PatternStore;
 
 import java.util.Objects;
